mitosheet: remove all widget infrastructure, move to mito_backend
First, I removed any notion of an ipywidget infrastructure from the frontend
and backend. This means we now just render using a `display()` call from
IPython.

As a result of this change, I had to do some CSS cleanup as well.

We also now use esbuild, which is _much_ faster than webpack.